+++ Gruntfile.js
@@ -131,10 +131,22 @@ module.exports = function (grunt) {
 
 		bump: {
 			options: {
-				files: ['package.json', 'bower.json'],
+				files: [
+					'package.json',
+					'bower.json',
+				],
 				updateConfigs: ['pkg'],
-				commitFiles: ['-a'],
-				pushTo: 'origin'
+				commitFiles: [
+					'package.json',
+					'bower.json',
+					'CHANGELOG.md',
+				],
+				commitMessage: 'Release v%VERSION%',
+				createTag: true,
+				tagName: 'v%VERSION%',
+				tagMessage: 'Release v%VERSION%',
+				push: true,
+				pushTo: 'origin',
 			}
 		},
 
@@ -226,6 +238,35 @@ module.exports = function (grunt) {
 			},
 		},
 
+		compress: {
+			installPackage: {
+				options: {
+					archive: '<%= pkg.name %>-<%= pkg.version %>.zip',
+					mode: 'zip',
+					pretty: true,
+				},
+				files: [
+					{
+						src: [
+							'inc/*',
+							'css/*',
+							'fonts/*',
+							'img/**',
+							'js/*',
+							'languages/*',
+							'template-parts/*',
+							'*.php',
+							'*.md',
+							'style.css',
+							'screenshot.png',
+							'readme.txt',
+						],
+						dest: '<%= pkg.name%>/',
+					},
+				]
+			}
+		}
+
 	});
 
 	grunt.registerTask('default', ['less', 'cssmin', 'copy']);
@@ -241,18 +282,24 @@ module.exports = function (grunt) {
 			'imagemin',
 			'copy',
 			'clean',
-			'conventionalChangelog'
-			//'bump-commit'
+			'conventionalChangelog',
+			'bump-commit',
+			'compress',
 		]);
 	});
 
 	grunt.registerTask('deploy', 'Deploy files to server over FTP.', function(account) {
-		if (account !== '') {
-			return grunt.task.run(['ftp-deploy:' + account])
+		if (typeof account !== 'undefined') {
+			return grunt.task.run(['ftp-deploy:' + account]);
 		} else {
-			return grunt.task.run(['ftp-deploy']);
+			return grunt.task.run([
+				'ftp-deploy:oddil',
+				'ftp-deploy:piskoviste',
+				'ftp-deploy:dobryweb',
+				'ftp-deploy:navod',
+			]);
 		}
 	});
 
-	grunt.registerTask('dev', 				['build', 'watch']);
-}
+	grunt.registerTask('dev', ['build', 'watch']);
+};